requirements.txt
pandas~=2.3.2
beautifulsoup4~=4.13.4
requests~=2.32.4
python-dotenv~=1.1.1